split route folder from controllers and rearrangement each folder relates #54
