+++ server/controllers/admin/logout.js
