+++ server/controllers/index.js
@@ -1,9 +1,47 @@
-const router = require('express').Router();
+const getCohortsData = require('./getCohortsData');
+const getSpecificCohort = require('./getSpecificCohort');
+const getProjectsData = require('./getProjectsByType');
+const getProjectData = require('./getProjectById');
+const getStudentProject = require('./getStudentProject');
+const getAlumniCohort = require('./getAlumniCohort');
+const getAlumni = require('./getAlumni');
+const getStudentById = require('./getStudentById');
+const getCohortProjects = require('./getCohortProjects');
+const {
+  addCohort,
+  addProject,
+  addStudentData,
+  deleteCohort,
+  deleteProjectData,
+  deleteStudent,
+  editCohort,
+  editProject,
+  login,
+  logout,
+  putStudentData,
+  getStats,
+} = require('./admin');
 
-const user = require('./routes/user');
-const admin = require('./routes/admin');
-
-router.use(user);
-router.use(admin);
-
-module.exports = router;
+module.exports = {
+  getCohortsData,
+  getSpecificCohort,
+  getProjectsData,
+  getProjectData,
+  getStudentProject,
+  getAlumniCohort,
+  getAlumni,
+  getStudentById,
+  addCohort,
+  addProject,
+  addStudentData,
+  deleteCohort,
+  deleteProjectData,
+  deleteStudent,
+  editCohort,
+  editProject,
+  getCohortProjects,
+  login,
+  logout,
+  putStudentData,
+  getStats,
+};
