--- server/controllers/routes/admin/cohort/index.js
@@ -1,9 +1,0 @@
-const addCohort = require('./addCohort');
-const deleteCohort = require('./deleteCohort');
-const editCohort = require('./editCohort');
-
-module.exports = {
-  addCohort,
-  deleteCohort,
-  editCohort,
-};
