--- server/controllers/routes/admin/project/index.js
@@ -1,11 +1,0 @@
-const addProject = require('./addProject');
-const getCohortProjects = require('./getCohortProjects');
-const editProject = require('./editProject');
-const deleteProjectData = require('./deleteProjectData');
-
-module.exports = {
-  addProject,
-  getCohortProjects,
-  editProject,
-  deleteProjectData,
-};
