--- server/controllers/routes/admin/student/index.js
@@ -1,9 +1,0 @@
-const deleteStudent = require('./deleteStudent');
-const putStudentData = require('./putStudentData');
-const addStudentData = require('./addStudentData');
-
-module.exports = {
-  deleteStudent,
-  addStudentData,
-  putStudentData,
-};
