--- server/controllers/routes/user/cohort/index.js
@@ -1,7 +1,0 @@
-const getCohortsData = require('./getCohortsData');
-const getSpecificCohort = require('./getSpecificCohort');
-
-module.exports = {
-  getCohortsData,
-  getSpecificCohort,
-};
