--- server/controllers/routes/user/project/index.js
@@ -1,9 +1,0 @@
-const getProjectsData = require('./getProjectsByType');
-const getProjectData = require('./getProjectById');
-const getStudentProject = require('./getStudentProject');
-
-module.exports = {
-  getProjectsData,
-  getProjectData,
-  getStudentProject,
-};
