--- server/controllers/routes/user/student/index.js
@@ -1,9 +1,0 @@
-const getAlumniCohort = require('./getAlumniCohort');
-const getAlumni = require('./getAlumni');
-const getStudentById = require('./getStudentById');
-
-module.exports = {
-  getAlumniCohort,
-  getAlumni,
-  getStudentById,
-};
